+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## In progress
 
 * Initial support for paging. Documented in the API guide, particularly the requiredRecords manifest entry.
+* Query parameters now broken out into a `params` property in the manifest.
 
 ## [0.3.0](https://github.com/folio-org/stripes-connect/tree/v0.3.0) (2017-04-06)
 [Full Changelog](https://github.com/folio-org/stripes-connect/compare/v0.2.0...v0.3.0)